requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.8.0
google-generativeai==0.3.2
unidecode==1.3.7
numpy==1.26.4
flask==2.3.2
flask-cors==4.0.0